Use esp-2021r2-patch3-8.4.0 for riscv builds only; use patch5 for all others
As part of updating esp-idf in earlier commits, the toolchain was updated to esp-2021r2-patch5-8.4.0
But this broke the ESP32-C3 (riscv32) builds. This commit uses patch3 for riscv builds and patch5
for all other builds.

Thanks to @Neradoc for noting that esp-idf/tools/tools.json file could support mixed versions.
I also tried to update tools/toolchain_versions.mk to support mixed versions based on architecture.
But this .mk file is actually read "by hand" in tools/cmake/crosstool_version_check.cmake to
extract version information, and that code assumes it is simple assignments without conditionals.
So instead I disabled the fatal error when there is a version mismatch, and changed it to a warning for
now.

We still need to figure out the actual reason why advancing the toolchain causes riscv builds
to boot loop.